Ensure data handling is done within process dataargs Removes unused dead code after adding the new framework and refactors some test cases and files.
