+++ tests/utils/test_preprocessing_utils.py
@@ -1,8 +1,8 @@
 # Third Party
 from datasets import Dataset
-from datasets.exceptions import DatasetGenerationError
 from transformers import AutoTokenizer, DataCollatorForSeq2Seq
 from trl import DataCollatorForCompletionOnlyLM
+import datasets
 import pytest
 
 # First Party
@@ -19,17 +19,12 @@ from tests.artifacts.testdata import (
 
 # Local
 from tuning.config import configs
-from tuning.data.setup_dataprocessor import process_dataargs
-from tuning.utils.preprocessing_utils import (
+from tuning.data.data_preprocessing_utils import (
     combine_sequence,
-    format_dataset,
     get_data_collator,
-    get_formatted_dataset_with_single_sequence,
-    get_preprocessed_dataset,
-    is_pretokenized_dataset,
-    load_hf_dataset_from_file,
     validate_data_args,
 )
+from tuning.data.setup_dataprocessor import is_pretokenized_dataset, process_dataargs
 
 
 @pytest.mark.parametrize(
@@ -66,91 +61,28 @@ def test_combine_sequence_adds_eos(input_element, output_element, expected_res):
     assert comb_seq == expected_res
 
 
-# Tests for loading the dataset from disk
 @pytest.mark.parametrize(
-    "dataset_path",
-    [TWITTER_COMPLAINTS_DATA_JSONL, TWITTER_COMPLAINTS_DATA_JSON],
-)
-def test_load_hf_dataset_from_file(dataset_path):
-    input_field_name = "Tweet text"
-    output_field_name = "text_label"
-    data = load_hf_dataset_from_file(
-        dataset_path,
-        input_field_name=input_field_name,
-        output_field_name=output_field_name,
-    )
-    # Our dataset should contain dicts that contain the input / output field name types
-    next_data = next(iter(data))
-    assert input_field_name in next_data
-    assert output_field_name in next_data
-
-
-def test_load_hf_dataset_from_jsonl_file_wrong_keys():
-    """Ensure that we explode if the keys are not in the jsonl file."""
-    with pytest.raises(DatasetGenerationError):
-        load_hf_dataset_from_file(
-            TWITTER_COMPLAINTS_DATA_JSONL,
-            input_field_name="foo",
-            output_field_name="bar",
-        )
-
-
-def test_load_hf_dataset_from_malformatted_data():
-    """Ensure that we explode if the data is not properly formatted."""
-    # NOTE: The actual keys don't matter here
-    with pytest.raises(DatasetGenerationError):
-        load_hf_dataset_from_file(
-            MALFORMATTED_DATA, input_field_name="foo", output_field_name="bar"
-        )
-
-
-def test_load_hf_dataset_from_jsonl_file_duplicate_keys():
-    """Ensure we cannot have the same key for input / output."""
-    with pytest.raises(ValueError):
-        load_hf_dataset_from_file(
-            TWITTER_COMPLAINTS_DATA_JSONL,
-            input_field_name="Tweet text",
-            output_field_name="Tweet text",
-        )
-
-
-# Tests for custom masking / preprocessing logic
-@pytest.mark.parametrize(
-    "dataset_path, max_sequence_length",
+    "data, result",
     [
-        (TWITTER_COMPLAINTS_DATA_JSONL, 1),
-        (TWITTER_COMPLAINTS_DATA_JSONL, 10),
-        (TWITTER_COMPLAINTS_DATA_JSONL, 100),
-        (TWITTER_COMPLAINTS_DATA_JSONL, 1000),
-        (TWITTER_COMPLAINTS_DATA_JSON, 1),
-        (TWITTER_COMPLAINTS_DATA_JSON, 10),
-        (TWITTER_COMPLAINTS_DATA_JSON, 100),
-        (TWITTER_COMPLAINTS_DATA_JSON, 1000),
+        (TWITTER_COMPLAINTS_DATA_JSONL, False),
+        (
+            Dataset.from_list(
+                [
+                    {
+                        "input_ids": [9437, 29, 210],
+                        "attention_mask": [1, 1, 1],
+                        "labels": [1, 20, 30],
+                    }
+                ]
+            ),
+            True,
+        ),
     ],
 )
-def test_get_preprocessed_dataset(dataset_path, max_sequence_length):
-    """Ensure we can handle preprocessed datasets with different max_sequence_lengths
-    to ensure proper tokenization and truncation.
-    """
-    tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)
-    preprocessed_data = get_preprocessed_dataset(
-        data_path=dataset_path,
-        tokenizer=tokenizer,
-        max_sequence_length=max_sequence_length,
-        input_field_name="Tweet text",
-        output_field_name="text_label",
-    )
-    for tok_res in preprocessed_data:
-        # Since the padding is left to the collator, there should be no 0s in the attention mask yet
-        assert sum(tok_res["attention_mask"]) == len(tok_res["attention_mask"])
-        # If the source text isn't empty, we start with masked inputs
-        assert tok_res["labels"][0] == -100
-        # All keys in the produced record must be the same length
-        key_lengths = {len(tok_res[k]) for k in tok_res.keys()}
-        assert len(key_lengths) == 1
-        # And also that length should be less than or equal to the max length depending on if we
-        # are going up to / over the max size and truncating - padding is handled separately
-        assert key_lengths.pop() <= max_sequence_length
+def test_is_pretokenized_data(data, result):
+    """Ensure that the correct collator type is fetched based on the data args"""
+
+    assert is_pretokenized_dataset(data=data) == result
 
 
 @pytest.mark.parametrize(
@@ -159,10 +91,10 @@ def test_get_preprocessed_dataset(dataset_path, max_sequence_length):
         (
             False,
             "\n### Label:",
-            load_hf_dataset_from_file(
-                TWITTER_COMPLAINTS_DATA_JSONL,
-                input_field_name="Tweet text",
-                output_field_name="text_label",
+            datasets.load_dataset(
+                "json",
+                data_files=TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSON,
+                split="train",
             ),
             1024,
             DataCollatorForCompletionOnlyLM,
@@ -196,33 +128,10 @@ def test_get_data_collator(
         packing,
         response_template,
         AutoTokenizer.from_pretrained(MODEL_NAME),
-        formatted_train_dataset,
+        is_pretokenized_dataset(formatted_train_dataset),
         max_seq_length,
     )
     assert isinstance(collator, expected_collator)
-
-
-@pytest.mark.parametrize(
-    "data, result",
-    [
-        (TWITTER_COMPLAINTS_DATA_JSONL, False),
-        (
-            Dataset.from_list(
-                [
-                    {
-                        "input_ids": [9437, 29, 210],
-                        "attention_mask": [1, 1, 1],
-                        "labels": [1, 20, 30],
-                    }
-                ]
-            ),
-            True,
-        ),
-    ],
-)
-def test_is_pretokenized_data(data, result):
-    """Ensure that the correct collator type is fetched based on the data args"""
-    assert is_pretokenized_dataset(data=data) == result
 
 
 # Tests for validating data args
@@ -314,7 +223,11 @@ def test_is_pretokenized_data(data, result):
 def test_validate_args(data_args, packing):
     """Ensure that respective errors are thrown for incorrect data arguments"""
     with pytest.raises(ValueError):
-        validate_data_args(data_args, packing)
+        is_traindata_tokenized = is_pretokenized_dataset(data_args.training_data_path)
+        is_evaldata_tokenized = is_pretokenized_dataset(data_args.validation_data_path)
+        validate_data_args(
+            data_args, packing, is_traindata_tokenized, is_evaldata_tokenized
+        )
 
 
 @pytest.mark.parametrize(
@@ -339,35 +252,11 @@ def test_validate_args(data_args, packing):
 )
 def test_validate_args_pretokenized(data_args, packing):
     """Ensure that supported data args do not error out when passing pretokenized datasets"""
-    validate_data_args(data_args, packing)
-
-
-@pytest.mark.parametrize(
-    "data_path, dataset_text_field, data_formatter_template",
-    [
-        (TWITTER_COMPLAINTS_DATA_JSON, "output", None),
-        (TWITTER_COMPLAINTS_DATA_JSONL, "output", None),
-        (
-            TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSON,
-            "formatted_field",
-            "### Text:{{input}} \n\n### Label: {{output}}",
-        ),
-        (
-            TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSONL,
-            "formatted_field",
-            "### Text:{{input}} \n\n### Label: {{output}}",
-        ),
-    ],
-)
-def test_get_formatted_dataset_with_single_sequence(
-    data_path, dataset_text_field, data_formatter_template
-):
-    tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)
-    formatted_dataset = get_formatted_dataset_with_single_sequence(
-        data_path, dataset_text_field, tokenizer, data_formatter_template
+    is_traindata_tokenized = is_pretokenized_dataset(data_args.training_data_path)
+    is_evaldata_tokenized = is_pretokenized_dataset(data_args.validation_data_path)
+    validate_data_args(
+        data_args, packing, is_traindata_tokenized, is_evaldata_tokenized
     )
-    assert isinstance(formatted_dataset, Dataset)
-    assert dataset_text_field in formatted_dataset.column_names
 
 
 @pytest.mark.parametrize(
@@ -396,8 +285,8 @@ def test_get_formatted_dataset_with_single_sequence(
             configs.DataArguments(
                 training_data_path=TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSON,
                 validation_data_path=TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSON,
-                dataset_text_field="formatted_field",
                 data_formatter_template="### Text:{{input}} \n\n### Label: {{output}}",
+                response_template="\n### Label:",
             )
         ),
         # data formatter template with input/output JSONL
@@ -405,80 +294,8 @@ def test_get_formatted_dataset_with_single_sequence(
             configs.DataArguments(
                 training_data_path=TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSONL,
                 validation_data_path=TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSONL,
-                dataset_text_field="formatted_field",
                 data_formatter_template="### Text:{{input}} \n\n### Label: {{output}}",
-            )
-        ),
-        # input/output JSON with masking on input
-        (
-            configs.DataArguments(
-                training_data_path=TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSON,
-                validation_data_path=TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSON,
-            )
-        ),
-        # input/output JSONL with masking on input
-        (
-            configs.DataArguments(
-                training_data_path=TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSONL,
-                validation_data_path=TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSONL,
-            )
-        ),
-    ],
-)
-def test_format_dataset(data_args):
-    """Ensure that the train/eval data are properly formatted based on the data args / text field"""
-    tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)
-    train_set, eval_set, dataset_text_field = format_dataset(
-        data_args, tokenizer, max_seq_length=1024
-    )
-    assert isinstance(train_set, Dataset)
-    assert isinstance(eval_set, Dataset)
-    if dataset_text_field is None:
-        column_names = set(["input_ids", "attention_mask", "labels"])
-        assert set(eval_set.column_names) == column_names
-        assert set(train_set.column_names) == column_names
-    else:
-        assert dataset_text_field in train_set.column_names
-        assert dataset_text_field in eval_set.column_names
-
-
-@pytest.mark.parametrize(
-    "data_args",
-    [
-        # single sequence JSON and response template
-        (
-            configs.DataArguments(
-                training_data_path=TWITTER_COMPLAINTS_DATA_JSON,
-                validation_data_path=TWITTER_COMPLAINTS_DATA_JSON,
-                dataset_text_field="output",
                 response_template="\n### Label:",
-            )
-        ),
-        # single sequence JSONL and response template
-        (
-            configs.DataArguments(
-                training_data_path=TWITTER_COMPLAINTS_DATA_JSONL,
-                validation_data_path=TWITTER_COMPLAINTS_DATA_JSONL,
-                dataset_text_field="output",
-                response_template="\n### Label:",
-            )
-        ),
-        # data formatter template with input/output JSON
-        (
-            configs.DataArguments(
-                training_data_path=TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSON,
-                validation_data_path=TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSON,
-                dataset_text_field="formatted_field",
-                data_formatter_template="### Text:{{input}} \n\n### Label: {{output}}",
-            )
-        ),
-        # data formatter template with input/output JSONL
-        (
-            configs.DataArguments(
-                training_data_path=TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSONL,
-                validation_data_path=TWITTER_COMPLAINTS_DATA_INPUT_OUTPUT_JSONL,
-                dataset_text_field="formatted_field",
-                data_formatter_template="### Text:{{input}} \n\n### Label: {{output}}",
             )
         ),
         # input/output JSON with masking on input
@@ -500,8 +317,13 @@ def test_format_dataset(data_args):
 def test_process_dataargs(data_args):
     """Ensure that the train/eval data are properly formatted based on the data args / text field"""
     tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)
-    train_set, eval_set, dataset_text_field = process_dataargs(
-        data_args, tokenizer, max_seq_length=1024
+    TRAIN_ARGS = configs.TrainingArguments(
+        packing=False,
+        max_seq_length=1024,
+        output_dir="tmp",  # Not needed but positional
+    )
+    (train_set, eval_set, dataset_text_field, _, _, _) = process_dataargs(
+        data_args, tokenizer, TRAIN_ARGS
     )
     assert isinstance(train_set, Dataset)
     assert isinstance(eval_set, Dataset)
@@ -512,49 +334,6 @@ def test_process_dataargs(data_args):
     else:
         assert dataset_text_field in train_set.column_names
         assert dataset_text_field in eval_set.column_names
-
-
-@pytest.mark.parametrize(
-    "data_args",
-    [
-        # JSON pretokenized train and validation datasets
-        (
-            configs.DataArguments(
-                training_data_path=TWITTER_COMPLAINTS_TOKENIZED_JSON,
-                validation_data_path=TWITTER_COMPLAINTS_TOKENIZED_JSON,
-            )
-        ),
-        # JSONL pretokenized train and validation datasets
-        (
-            configs.DataArguments(
-                training_data_path=TWITTER_COMPLAINTS_TOKENIZED_JSONL,
-                validation_data_path=TWITTER_COMPLAINTS_TOKENIZED_JSONL,
-            )
-        ),
-        # JSON pretokenized train datasets
-        (
-            configs.DataArguments(
-                training_data_path=TWITTER_COMPLAINTS_TOKENIZED_JSON,
-            )
-        ),
-        # JSONL pretokenized train datasets
-        (
-            configs.DataArguments(
-                training_data_path=TWITTER_COMPLAINTS_TOKENIZED_JSONL,
-            )
-        ),
-    ],
-)
-def test_format_dataset_pretokenized(data_args):
-    """Ensure that pretokenized datasets are loaded and returned as is"""
-    train_set, eval_set, _ = format_dataset(data_args, None, max_seq_length=1024)
-    assert isinstance(train_set, Dataset)
-    if eval_set:
-        assert isinstance(eval_set, Dataset)
-
-    assert set(["input_ids", "labels"]).issubset(set(train_set.column_names))
-    if eval_set:
-        assert set(["input_ids", "labels"]).issubset(set(eval_set.column_names))
 
 
 @pytest.mark.parametrize(
@@ -590,7 +369,15 @@ def test_format_dataset_pretokenized(data_args):
 )
 def test_process_dataargs_pretokenized(data_args):
     """Ensure that pretokenized datasets are loaded and returned as is"""
-    train_set, eval_set, _ = process_dataargs(data_args, None, max_seq_length=1024)
+    TRAIN_ARGS = configs.TrainingArguments(
+        packing=False,
+        max_seq_length=1024,
+        output_dir="tmp",  # Not needed but positional
+    )
+    tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)
+    (train_set, eval_set, _, _, _, _) = process_dataargs(
+        data_args, tokenizer, TRAIN_ARGS
+    )
     assert isinstance(train_set, Dataset)
     if eval_set:
         assert isinstance(eval_set, Dataset)
